update proposal_id of project to BIGINT

+++ migrations/20250626174153_project-add-proposal-id.up.sql
@@ -1,3 +1,3 @@
 -- Add up migration script here
 
-ALTER TABLE project ADD COLUMN IF NOT EXISTS proposal_id INT;
+ALTER TABLE project ADD COLUMN IF NOT EXISTS proposal_id BIGINT;
